moved out models out of their packages to main models' directory
