+++ project/alembic/env.py
@@ -5,7 +5,7 @@ from sqlalchemy import engine_from_config
 from sqlalchemy import pool
 from alembic import context
 
-from database.entities.base import Base
+from database.models.base import Base
 
 
 # this is the Alembic Config object, which provides
