--- project/database/entities/__init__.py
@@ -1,2 +1,0 @@
-from database.entities.dogs.breeds_models import Breed
-from database.entities.user.user_models import User
