+++ project/database/models/__init__.py
@@ -1,0 +1,2 @@
+from database.models.dog import Breed
+from database.models.user import User
